+++ packages/proto-signing/src/directethsecp256k1hdwallet.ts
@@ -1,0 +1,168 @@
+import { encodeEthSecp256k1Signature, rawEthSecp256k1PubkeyToRawAddress } from "@cosmjs/amino";
+import {
+  Bip39,
+  EnglishMnemonic,
+  HdPath,
+  keccak256,
+  Random,
+  Secp256k1,
+  Secp256k1Keypair,
+  Slip10,
+  Slip10Curve,
+  Slip10RawIndex,
+} from "@cosmjs/crypto";
+import { toBech32 } from "@cosmjs/encoding";
+import { SignDoc } from "cosmjs-types/cosmos/tx/v1beta1/tx";
+
+import { AccountData, DirectSignResponse, OfflineDirectSigner } from "./signer";
+import { makeSignBytes } from "./signing";
+
+interface AccountDataWithPrivkey extends AccountData {
+  readonly privkey: Uint8Array;
+}
+
+/**
+ * Derivation information required to derive a keypair and an address from a mnemonic.
+ */
+interface Secp256k1Derivation {
+  readonly hdPath: HdPath;
+  readonly prefix: string;
+}
+
+export interface DirectEthSecp256k1HdWalletOptions {
+  /** The password to use when deriving a BIP39 seed from a mnemonic. */
+  readonly bip39Password: string;
+  /** The BIP-32/SLIP-10 derivation paths. Defaults to the Cosmos Hub/ATOM path `m/44'/60'/0'/0/0`. */
+  readonly hdPaths: readonly HdPath[];
+  /** The bech32 address prefix (human readable part). Defaults to "cosmos". */
+  readonly prefix: string;
+}
+
+interface DirectEthSecp256k1HdWalletConstructorOptions extends Partial<DirectEthSecp256k1HdWalletOptions> {
+  readonly seed: Uint8Array;
+}
+
+const defaultOptions: DirectEthSecp256k1HdWalletOptions = {
+  bip39Password: "",
+  hdPaths: [
+    [
+      Slip10RawIndex.hardened(44),
+      Slip10RawIndex.hardened(60),
+      Slip10RawIndex.hardened(0),
+      Slip10RawIndex.normal(0),
+      Slip10RawIndex.normal(0),
+    ] as HdPath,
+  ],
+  prefix: "cosmos",
+};
+
+/** A wallet for protobuf based signing using SIGN_MODE_DIRECT */
+export class DirectEthSecp256k1HdWallet implements OfflineDirectSigner {
+  /**
+   * Restores a wallet from the given BIP39 mnemonic.
+   *
+   * @param mnemonic Any valid English mnemonic.
+   * @param options An optional `DirectEthSecp256k1HdWalletOptions` object optionally containing a bip39Password, hdPaths, and prefix.
+   */
+  public static async fromMnemonic(
+    mnemonic: string,
+    options: Partial<DirectEthSecp256k1HdWalletOptions> = {},
+  ): Promise<DirectEthSecp256k1HdWallet> {
+    const mnemonicChecked = new EnglishMnemonic(mnemonic);
+    const seed = await Bip39.mnemonicToSeed(mnemonicChecked, options.bip39Password);
+    return new DirectEthSecp256k1HdWallet(mnemonicChecked, {
+      ...options,
+      seed: seed,
+    });
+  }
+
+  /**
+   * Generates a new wallet with a BIP39 mnemonic of the given length.
+   *
+   * @param length The number of words in the mnemonic (12, 15, 18, 21 or 24).
+   * @param options An optional `DirectSecp256k1HdWalletOptions` object optionally containing a bip39Password, hdPaths, and prefix.
+   */
+  public static async generate(
+    length: 12 | 15 | 18 | 21 | 24 = 12,
+    options: Partial<DirectEthSecp256k1HdWalletOptions> = {},
+  ): Promise<DirectEthSecp256k1HdWallet> {
+    const entropyLength = 4 * Math.floor((11 * length) / 33);
+    const entropy = Random.getBytes(entropyLength);
+    const mnemonic = Bip39.encode(entropy);
+    return DirectEthSecp256k1HdWallet.fromMnemonic(mnemonic.toString(), options);
+  }
+
+  /** Base secret */
+  private readonly secret: EnglishMnemonic;
+  /** BIP39 seed */
+  private readonly seed: Uint8Array;
+  /** Derivation instructions */
+  private readonly accounts: readonly Secp256k1Derivation[];
+
+  protected constructor(mnemonic: EnglishMnemonic, options: DirectEthSecp256k1HdWalletConstructorOptions) {
+    const prefix = options.prefix ?? defaultOptions.prefix;
+    const hdPaths = options.hdPaths ?? defaultOptions.hdPaths;
+    this.secret = mnemonic;
+    this.seed = options.seed;
+    this.accounts = hdPaths.map((hdPath) => ({
+      hdPath: hdPath,
+      prefix: prefix,
+    }));
+  }
+
+  public get mnemonic(): string {
+    return this.secret.toString();
+  }
+
+  public async getAccounts(): Promise<readonly AccountData[]> {
+    const accountsWithPrivkeys = await this.getAccountsWithPrivkeys();
+    return accountsWithPrivkeys.map(({ algo, pubkey, address }) => ({
+      algo: algo,
+      pubkey: pubkey,
+      address: address,
+    }));
+  }
+
+  public async signDirect(signerAddress: string, signDoc: SignDoc): Promise<DirectSignResponse> {
+    const accounts = await this.getAccountsWithPrivkeys();
+    const account = accounts.find(({ address }) => address === signerAddress);
+    if (account === undefined) {
+      throw new Error(`Address ${signerAddress} not found in wallet`);
+    }
+    const { privkey, pubkey } = account;
+    const signBytes = makeSignBytes(signDoc);
+    const hashedMessage = keccak256(signBytes);
+
+    const signature = Secp256k1.createSignature(hashedMessage, privkey);
+    const signatureBytes = new Uint8Array([...signature.r(32), ...signature.s(32)]);
+    const stdSignature = encodeEthSecp256k1Signature(pubkey, signatureBytes);
+    return {
+      signed: signDoc,
+      signature: stdSignature,
+    };
+  }
+
+  private async getKeyPair(hdPath: HdPath): Promise<Secp256k1Keypair> {
+    const { privkey } = Slip10.derivePath(Slip10Curve.Secp256k1, this.seed, hdPath);
+    const { pubkey } = Secp256k1.makeKeypair(privkey);
+    return {
+      privkey: privkey,
+      pubkey: Secp256k1.compressPubkey(pubkey),
+    };
+  }
+
+  private async getAccountsWithPrivkeys(): Promise<readonly AccountDataWithPrivkey[]> {
+    return Promise.all(
+      this.accounts.map(async ({ hdPath, prefix }) => {
+        const { privkey, pubkey } = await this.getKeyPair(hdPath);
+        const address = toBech32(prefix, rawEthSecp256k1PubkeyToRawAddress(pubkey));
+        return {
+          algo: "eth_secp256k1" as const,
+          privkey: privkey,
+          pubkey: pubkey,
+          address: address,
+        };
+      }),
+    );
+  }
+}