+++ CHANGELOG.md
@@ -23,11 +23,13 @@ and this project adheres to
   price from the chain. ([#1926])
 - @cosmjs/cosmwasm-stargate: Add the ability to specify a custom account parser
   for `CosmWasmClient`. ([#1928])
+- Add support for Cosmos EVM key handling and signing. ([#1932])
 
 [#1883]: https://github.com/cosmos/cosmjs/issues/1883
 [#1916]: https://github.com/cosmos/cosmjs/pull/1916
 [#1926]: https://github.com/cosmos/cosmjs/pull/1926
 [#1928]: https://github.com/cosmos/cosmjs/pull/1928
+[#1932]: https://github.com/cosmos/cosmjs/pull/1932
 
 ### Changed